backend/apis/langchain-api/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "app" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.poetry]
name = "langchain-api"
version = "0.0.1"
description = "Langchain API"
authors = [
  "Vuong Ngo <vuongngo.pd@gmail.com>"
]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4"
uvicorn = "0.28.1"
fastapi = "0.110.0"
pydantic = "2.6.4"
langchain-core = "0.1.42"
langchain-openai = "0.1.0"
langchain-community = "^0.0.32"
langchain = "^0.1.16"
python-dotenv = "^1.0.1"
redis = "^5.0.3"
syntrac-sdk = "0.0.3"

[tool.poetry.group.dev.dependencies]
autopep8 = "2.0.4"
flake8 = "7.0.0"
pytest = "8.0.2"
pytest-sugar = "1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
